refactor(memory): introduce ConsolidationRound and extract consolidation.py
